+++ src/components/Slider/swiper_data.js
